[feat] fix error after code refactorying
