+++ launch/drive_prediction.launch
@@ -7,7 +7,7 @@
   <!-- <node pkg="parking" type="path_maker.py" name="parking" /> -->
 
   <!-- CONTROL_Pure Pursuit -->
-  <node pkg="motion_planner" type="rule_based_planner_avoid_obstacle.py" name="rule_based_planner" output="screen"/>
+  <node pkg="motion_planner" type="rule_based_planner_avoid_obstacle_pid.py" name="rule_based_planner" output="screen"/>
 
   <!-- GPS AND IMU SENSOR -->
   <node pkg="state_estimator" type="gpsimu_parser.py" name="GPS_IMU_parser" />
@@ -22,6 +22,7 @@
   <node pkg = "rviz_runner" type="tf_pub.py" name="tf" />
   <node pkg="rviz" type="rviz" name="rviz" args="-d $(find rviz_runner)/rviz/kcity_rviz.rviz"/>
 
+  <!-- <node pkg="control" type="lateral_controller_stanley.py" name="lateral_controller" output="screen"/> -->
 
   <!-- <node pkg="control" type="longitudinal_controller.py" name="longitudinal_controller" output="screen"/> -->
   <!-- <node pkg="control" type="experiment.py" name="experiment" output="screen"/> -->
